+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * `Ga4FormTracker` handles `selectedIndex` of `-1` ([PR #4803](https://github.com/alphagov/govuk_publishing_components/pull/4803))
 * Extra meta tags for `PageViewTracker` ([PR #4804](https://github.com/alphagov/govuk_publishing_components/pull/4804/))
+* Fix edge cases with YouTube link enhancement JS ([PR #4808](https://github.com/alphagov/govuk_publishing_components/pull/4808))
 
 ## 56.3.1